backend/requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
setuptools<58.0.0 
psycopg2
celery==5.3.1
django-celery-results==2.4.0
redis==5.0.0
django-cors-headers==4.2.0
django-environ==0.10.0
requests==2.31.0
ptvsd==4.3.2
pytest-django==4.5.2
pytest-factoryboy==2.3.1
flake8==5.0.4
isort==5.6.4
gevent==23.7.0